ru to double

+++ dbms/src/Flash/Executor/toRU.h
@@ -18,8 +18,7 @@
 
 namespace DB
 {
-// gjt todo change to double
-using RU = UInt64;
+using RU = double;
 
 // Convert cpu time nanoseconds to cpu time millisecond, and round up.
 UInt64 toCPUTimeMillisecond(UInt64 cpu_time_ns);
